Use ⌽¯4↑ instead of 4↑⌽
This is one more character and perhaps not as æsthetic, but reversing
that whole array seems like a waste just to get to the last four bits.
Plus, at this point 2⊸×⊸+˜´⌽ feels like a block itself to me, and having
the reverse baked into the previous manipulation might be annoying when
re-reading this part of the code.